Merge phase 3 into main

+++ core/src/system_stm32f4xx.c
@@ -61,7 +61,7 @@ void SystemInit(void)
      * SRAM, or if a bootloader hands off execution with VTOR already
      * pointing elsewhere.
      *--------------------------------------------------------------*/
-    SCB_VTOR = (uint32_t)g_pfnVectors;
+    SCB_VTOR = (uintptr_t)g_pfnVectors;
 }
 
 /*-----------------------------------------------------------------------*/
